ASSET-26 Apply testcase option
